Reworked executing puppet, added more output
Prior to this, two different types of execution were in use, one of
which was not cross platform. Calls to puppet were also done in multiple
places. After this change there is a single function that handles all
calls to puppet. This function uses the existing PATH and appends the
default location of Puppet while taking into account that there are
differences on Windows. Some initial debug output was also added that is
only shown if the `PUPPET_CRON_DEBUG` environment variable is set.