Parse cli arguments, allow setting fallback env
Prior to this, the fallback environment was hard coded to production as
were the arguments passed to puppet to run the agent. Now there is a
flag for overridding the fallback environment (-env) and one to enable
debugging output from puppet-runner itself (-debug). If there are any
arguments pased besides these flags they will be treated as a string and
override the defaults passed when running the agent.